Replaced short usage with byte at necessary places

+++ Applet/AndroidSEProviderLib/src/com/android/javacard/seprovider/KMAndroidSEProvider.java
@@ -41,17 +41,17 @@ import javacard.security.KeyAgreement;
 
 public class KMAndroidSEProvider implements KMSEProvider {
 
-  public static final short AES_GCM_TAG_LENGTH = 16;
-  public static final short AES_GCM_NONCE_LENGTH = 12;
+  public static final byte AES_GCM_TAG_LENGTH = 16;
+  public static final byte AES_GCM_NONCE_LENGTH = 12;
   public static final byte KEYSIZE_128_OFFSET = 0x00;
   public static final byte KEYSIZE_256_OFFSET = 0x01;
   public static final short TMP_ARRAY_SIZE = 300;
   private static final short RSA_KEY_SIZE = 256;
   public static final short CERT_CHAIN_MAX_SIZE = 2500;//First 2 bytes for length.
-  public static final short SHARED_SECRET_KEY_SIZE = 32;
+  public static final byte SHARED_SECRET_KEY_SIZE = 32;
   public static final byte POWER_RESET_FALSE = (byte) 0xAA;
   public static final byte POWER_RESET_TRUE = (byte) 0x00;
-  private static final short COMPUTED_HMAC_KEY_SIZE = 32;
+  private static final byte COMPUTED_HMAC_KEY_SIZE = 32;
   private static byte[] CMAC_KDF_CONSTANT_L;
   private static byte[] CMAC_KDF_CONSTANT_ZERO;
